Renamed the package fi.fmi.avi.model.SpaceWeatherAdvisory to fi.fmi.avi.model.swx
