+++ src/main/java/fi/fmi/avi/converter/json/SpaceWeatherAdvisoryJSONParser.java
@@ -3,8 +3,8 @@ package fi.fmi.avi.converter.json;
 import fi.fmi.avi.converter.AviMessageSpecificConverter;
 import fi.fmi.avi.converter.ConversionHints;
 import fi.fmi.avi.converter.ConversionResult;
-import fi.fmi.avi.model.SpaceWeatherAdvisory.SpaceWeatherAdvisory;
-import fi.fmi.avi.model.SpaceWeatherAdvisory.immutable.SpaceWeatherAdvisoryImpl;
+import fi.fmi.avi.model.swx.SpaceWeatherAdvisory;
+import fi.fmi.avi.model.swx.immutable.SpaceWeatherAdvisoryImpl;
 
 public class SpaceWeatherAdvisoryJSONParser extends AbstractJSONParser implements AviMessageSpecificConverter<String, SpaceWeatherAdvisory>  {
     @Override
