+++ src/main/java/fi/fmi/avi/model/swx/SpaceWeatherRegion.java
@@ -1,4 +1,4 @@
-package fi.fmi.avi.model.SpaceWeatherAdvisory;
+package fi.fmi.avi.model.swx;
 
 import java.util.Optional;
