+++ src/main/java/fi/fmi/avi/model/swx/immutable/NextAdvisoryImpl.java
@@ -1,4 +1,4 @@
-package fi.fmi.avi.model.SpaceWeatherAdvisory.immutable;
+package fi.fmi.avi.model.swx.immutable;
 
 import java.io.Serializable;
 
@@ -7,7 +7,7 @@ import org.inferred.freebuilder.FreeBuilder;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 
-import fi.fmi.avi.model.SpaceWeatherAdvisory.NextAdvisory;
+import fi.fmi.avi.model.swx.NextAdvisory;
 
 @FreeBuilder
 @JsonDeserialize(builder = NextAdvisoryImpl.Builder.class)
